added the ability to send the friend request notification

+++ server-Chat-App/src/main/java/org/server/chatapp/dao/dao/Dao.java
@@ -7,6 +7,6 @@ public interface Dao<T> {
     T get(long id);
     List<T> getAll();
     int update(T t);
-    int insert(T t);
+    long insert(T t);
     int delete(T t);
 }
